feat: 친구 찾기 기능 추가
