+++ frontend/app/src/view/index.ts
@@ -2,3 +2,4 @@ export * from './FriendView'
 export * from './GameView'
 export * from './ChatView'
 export * from './ProfileView'
+export * from './FindFriendView'
